database: Move header chain prover related things to their own module.
